Improve plugin loading during auditing phase
This commit updates how plugins are downloaded to use the new option to
download via npm.

It also hijacks the require() function for all babel plugin modules so
that @babel/core is loaded using the global copy installed in the babel
build plugin. This is a blatant hack to ensure the peer dependency is
met for each babel plugin without having to install a bunch of copies of
babel core.